+++ CHANGES.md
@@ -3,6 +3,7 @@ Change Log
 
 ### 1.36 - 2017-08-01
 
+* Added ability to disable security on the InfoBox, which removes the sandbox property and allows HTML in the title
 * Fixed a bug where a Model's compressed textures were not being displayed. [#5596](https://github.com/AnalyticalGraphicsInc/cesium/pull/5596)
 * Fixed a bug where jsep was undefined when using webpack [#5593](https://github.com/AnalyticalGraphicsInc/cesium/issues/5593)
 * Fixed documentation for `OrthographicFrustum` [#5586](https://github.com/AnalyticalGraphicsInc/cesium/issues/5586)